rio_api

Module model

Data structures for RDF 1.1 and RDF-star concepts like IRI, literal or triples.

Re-exports

pub use crate::model::*;

Structs

BlankNode

An RDF blank node.

GeneralizedQuad

A generalized RDF triple in a RDF dataset.

NamedNode

An RDF IRI.

Quad

A RDF triple in a RDF dataset.

StrictRdfError

An error raised when generalized RDF cannot be converted to strict RDF.

Triple

A RDF triple.

Variable

A SPARQL variable.

Enums

GeneralizedTerm

A generalized RDF term.

GraphName

A restriction of Term that can be used in the graph name position.

Literal

An RDF literal.

Subject

A restriction of Term that can be used as the subject of an RDF triple.

Term

An RDF term.